+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## Not released:
 
 * Add `node.isCached()` method
+* Fix nested dragging bug
 
 ## [4.0.3][2019-08-08]